+++ backend/src/main/java/com/dev/backend/service/EstadoService.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Service;
 import com.dev.backend.entity.Estado;
 import com.dev.backend.repository.EstadoRepository;
 
+import lombok.NonNull;
+
 @Service
 public class EstadoService {
 
@@ -30,8 +32,10 @@ public class EstadoService {
         return estadoRepository.saveAndFlush(obj);
     }
 
-    public void excluir(Long id) {
+    public void excluir(@NonNull Long id) {
         Estado obj = estadoRepository.findById(id).get();
+
+        if(obj != null)
         estadoRepository.delete(obj);
     }
 }
