+++ backend/src/main/java/com/dev/backend/service/ProdutoService.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Service;
 import com.dev.backend.entity.Produto;
 import com.dev.backend.repository.ProdutoRepository;
 
+import lombok.NonNull;
+
 @Service
 public class ProdutoService {
 
@@ -30,8 +32,11 @@ public class ProdutoService {
         return produtoRepository.saveAndFlush(obj);
     }
 
-    public void excluir(Long id) {
-        Produto obj = produtoRepository.findById(id).get();
-        produtoRepository.delete(obj);
+    public void excluir(@NonNull Long id) {
+
+            Produto obj = produtoRepository.findById(id).get();
+
+            if (obj != null) 
+                produtoRepository.delete(obj);
     }
 }
